requirement.txt
numpy==1.26.4
pandas==2.2.2
scipy==1.11.4
shapely==1.8.5
matplotlib==3.5.2
opencv-python==4.8.1.78
prettytable==3.7.0
pyquaternion==0.9.9
nuscenes-devkit==1.1.10
yapf==0.40.2
tensorboard==2.17.0
motmetrics==1.4.0
scikit-learn==1.3.2
tqdm==4.66.4
pillow==10.3.0
einops==0.7.0
pycocotools==2.0.7
urllib3==1.26.18
psutil==5.9.8
mmengine==0.10.4
mmcv==2.1.0
mmdet==3.3.0
mmdet3d==1.4.0
flash-attn==2.6.3
